Parallel processing of replication records
This change uses multiple threads for processing replication messages
arriving via a logical replication stream from Postgres master before
they are submitted to Kafka connect for further processing and flushing
to Kafka.

Life of a Logical Replication Record with Debezium:
1. WAL record is written by Postgres master in one of the threads
serving user queries.
2. WAL sender thread (which is serving records to a logical replication
client) reads the WAL entry, runs it through logical decoding to
reconstruct tuple metadata (columns changed, type info, row identity,
etc), and writes to the replication output stream.
3. Debezium's postgres connector's task main thread (postgres connector
is single-threaded) reads the record from the JDBC connection's
replication stream.
4. The main thread then moves forward with processing the newly fetched
record, e.g. converts to Kafka connect types.
5. The main thread then enqueues it in a bounded-size blocking queue
(henceforth called KCQueue) before going on to fetch the next record.
6. KCQueue is used for serving records to Kafka connect when it polls
periodically.  Its max size is `max.queue.size`.  Another param,
`max.batch.size` controls the maximum number of records that Kafka
connect will drain from the queue in a single attempt.
7. Kafka connect dequeues the record from the queue and does its own
processing on these records (e.g. apply transforms, etc).
8. Kafka connect finally writes to Kafka using a Kafka producer (single
partition only for Postgres connector).
9. Connector acks back to the master that a record has been flushed so
the master can update the `flush_lsn` for this slot.

This change optimizes step 4 where the connector's main thread, having
read a message from the source, processes it and writes to KCQueue from
where Kafka connect will pick it up.

The basic idea is to introduce a new queue, called the "Receive Queue"
between receipt of a logical replication message from Postgres and
processing it using a ThreadPoolExecutor.  The main thread reading from
the source logical replication stream adds each incoming record to
this queue and promptly goes back to reading from source.  In the
background, each of the thread pool's threads continuously read from
this queue, process them, and write the results to the KCQ.

Although the performance impact is immediate, for correctness (and
stability) following additional changes were needed:

1. The order of events received from source must match the order in
which they are served to Kafka connect.  This requires that the
processing threads ultimately serialize before outputing records.  We
went through a few different ideas on how to best achieve this.  This
was also made complicated by the fact that several of the intervening
classes that we wished to modify and extend are not postgres specific.
Ultimately, we ended up going with a (somewhat ugly and
abstraction-defeating) design which uses ThreadLocal state to buffer a
batch of processed records in a processing thread before outputing them.
This implementation is in BufferingChangeEventQueue.

2. The old code had several assumptions about monotonicity of LSN being
processed.  Since it was all happening in a single thread, in-order, the
LSN was being tracked in a single "global-like" variable.  Violating
this monotonicity assumption with parallelism broke this constraint and
resulted in random exceptions.  We handled this by replacing the global
LSN state with per-message LSN which is passed to the various processing
threads as params.  This can be seen in PgOutputMessageDecoder,
AbstractMessageDecoder, and TransactionMessage classes.

The core implementation is spread across the following classes:
1. PostgresConnectorTask: The main connector implementation which
orchestrates everything.  Adds the receiveQueue and
BufferingChangeEventQueue.  Note that receiveQueue contains messages of
type RawReplicationMessage, whereas BufferingChangeEventQueue is simply
a wrapper around KCQueue and adds buffering functionality to it.
2. PostgresStreamingChangeEventSource: Main execution loop which runs
indefinitely.  This loop is now tweaked to enable background processing.
3. PostgresReplicationConnection: Main reader of the replication stream.
This reads messages, enqueues to receive queue, and hosts the thread
pool processing off the receive queue.

Observability:
- Previously KCQueue utilization was a good whitebox metric for
performance of this pipeline.  If this queue was mostly full, that would
indicate Kafka connect was not draining it fast enough.  If mostly
empty, that would mean reading from source was the bottleneck.
- With this change, we introduced monitoring for the receive queue as
well.  Our goal is to make sure that it's rarely full.  This is
implemented via changes to PostgresStreamingChangeEventSourceMetrics
and PostgresChangeEventSourceMetricsFactory.

Performance Benchmarks:
- Raw Postgres replication performance from RDS WAL: 160MBps.
- Upstream Debezium: 45MBps.
- This change: 140-150MBps.

Testing Done:
- Zero exceptions when reading 300GB of WAL data written by HammerDB
(tpc-c) benchmark across 20+ runs.

Testing Remaining:
- Add automated correctness testing which verifies ordering and ensures
there is no loss when compared to upstream Debezium.